+++ src/api.tsx
@@ -5,6 +5,7 @@ import { v4 as newUuid } from 'uuid';
 import sha256 from 'crypto-js/sha256';
 import Patient from './misc/patient';
 import { TestResult } from './misc/enum';
+import ITestResult from './misc/test-result';
 
 export const api = axios.create({
     baseURL: ''
@@ -12,7 +13,7 @@ export const api = axios.create({
 
 const TRYS = 2;
 
-export const usePostTestResult = (testResult: TestResult | undefined, processId: string, onSuccess?: () => void, onError?: (error: any) => void) => {
+export const usePostTestResult = (testResult: ITestResult | undefined, processId: string, onSuccess?: () => void, onError?: (error: any) => void) => {
     const { keycloak, initialized } = useKeycloak();
 
     React.useEffect(() => {
@@ -20,10 +21,7 @@ export const usePostTestResult = (testResult: TestResult | undefined, processId:
         if (testResult && processId) {
 
             const uri = '/api/quicktest/' + processId + '/testResult';
-            const body = JSON.stringify({
-                result: testResult
-            });
-
+            const body = JSON.stringify(testResult);
 
             const header = {
                 "Authorization": initialized ? `Bearer ${keycloak.token}` : "",
@@ -56,7 +54,6 @@ export const usePostPatient = (patient: Patient | undefined, processId: string, 
             const body = JSON.stringify({
                 confirmationCwa: patient.processingConsens,
                 insuranceBillStatus: patient.billStatus,
-                testBrandId: patient.testId,
                 lastName: patient.name,
                 firstName: patient.firstName,
                 email: patient.emailAddress,
