+++ src/misc/test-result.tsx
@@ -1,0 +1,7 @@
+import { TestResult } from './enum';
+
+export default interface ITestResult {
+    result: TestResult;
+    testId: string;
+    testBrandId?: string;
+}
